data-generator/requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.0.32
python-dateutil==2.8.2